latest C6/C2 wip  (#19344)
* Initial C6 support

* add C6 in tasmota_globals.h

* fix copy&paste errors

* Use updated platform (build 1481)

* use common safeboot flags (#19282) (#352)


* use env var lib_ignore

* safeboot lib_ignore env

* revert wrong change

* rm eink lib from standard used libs

* add c2-board 4M, maximum_size probably a bit off

* diffuse idea for 2MB with OTA and FS

* C2 and C6 additions

* C2 and C6 additions for commands.ino

* C2 and C6 additions for crash recorder

* C2 to support_esp

* some preps for C2

* add c2 to tasmota.h

* fix some typos and compile bugs

* fix 2 compiler warnings for Arduino 3.0

* fix another warning

* add correct pins to template

* pins for C2, maximum_size updated

* unclutter env:tasmota32c2-safeboot

* fix c6 env

* more pins for the C2, but not all

* pin 9 and 10 for c2/c6

* add ||

* prep pins for c6

* make c6 compile, needs testing

* fix copy-paste-error

* more correct pin support C6

* pin14 C6

* wrong warning for C6

* 10 and 11 are flash pins on the C6

* boolean voodoo

* fix EVERYTHING

* fix Onewire for Arduino 3.0

* enable OneWire

* finale fix for all platforms: OneWire.h

* better compatibility

* final fix onewire

* refactor Platformio Arduino 3.0 setup

* overlooked needed `;`

* refactor as requested

* fix NeoPixelBus compilation for C2 and C6

* fix Arduino30 for old working NeoPixel boards

* Try to switch of RMT LED when not SOC supported

* use latest IDF 5.1

* framework includes now "libsodium"

* prep NeoPixel RMT rewrite

* NeoPixel for C6

* fix Arduino 2.x

* really fix compile ??

* fix compiler warning TSL2591

* fix compiler warning TSL2591 _2

* fix compile warning TSL2561

* fix warning sensirion part 1

* fix ESP32 Arduino 3.0 compile (needs work check)

* Disable NeoPixel for C2 (no RMT support)

* forgotten to remove

* fix compiler warning

* prep 5.1, TasUpdater

* prep 5.1 SensirionRXFrame

* prep 5.1, SensiriomI2CCommunication.cpp

* prep5.1, last batch Sensirion

* prep5.1, silence compiler warning

* Silence C++ deprecated warning

* Align with branch C6-support (#359)

* prep5.1, fix berry_wire

* revert deprecated compiler flag

---------


* add i2c #defines

* enable more libs for ESP32

* refactor for esp-nimble-cpp

* add esp-nimble-cpp, needed for C2,C6

* fix dual use of IRAM_ATTR for i2sDmaISR

* silence compiler warning with cast

* prep5.1, new DAC API in Berry

* more libs enabled

* NimBLE changes

* fix wrong build flag for MI

* add NimBLE ignores

* add MI to env

* use latest framework with changed NimBLE

* add libesp32_div to C2/C6

* fix env konfig for NimBLE

* new esp-cpp-nimble for all

* prep5.1, silence compiler warning around IRAM_ATR

* prep5.1, i2s pre-work

* add c2 to lib esp-nimble-cpp

* prep5.1, intermediate refactoring

* prep5.1, intermediate refactoring

* fix chirp for IDF>5

* prep5.1, fix compilation error

* Update platformio_tasmota_core3_env.ini

* prep5.1, SPI fixes for C2/C6

* prep5.1, fix universal display

* enable uDisplay

* c3cdc display

* add odroid env

* rm obsolote option

* rm BLE from c3cdc

* fix SD Card (changed sdkconfig)

* cdc env, lvgl to all builds

* add cdc safeboot

* use framework build 1504

* prep5.1., intermediate restart

* forgot to save

* prep5.1, we love static cast

* latest framework build 1510

* fix xsns_74_lmt01

* ditch NeoPixel-I2S for 5.1

* prep5.1, begin RMT C6

* Ethernet Test

* forgot to upload

* invert timing in led-encoder

* build 1511 / IDF 5.1.0.230818

* compile NeopixelBus for C6

* re - uncomment

* re - uncomment the next

* enable Neopixel

* One more Neopixel try

* what a stupid typo

* refac Pio Arduino 3.0 setup

* updates ini's

* disable by default core 32 Arduino 3.0

* disable uDisplay for S3

* Update tasmota_configurations_ESP32.h

---------

+++ lib/libesp32_lvgl/lvgl/src/core/lv_obj_style.h
@@ -91,7 +91,7 @@ void lv_obj_remove_style(struct _lv_obj_t * obj, lv_style_t * style, lv_style_se
  */
 static inline void lv_obj_remove_style_all(struct _lv_obj_t * obj)
 {
-    lv_obj_remove_style(obj, NULL, LV_PART_ANY | LV_STATE_ANY);
+    lv_obj_remove_style(obj, NULL, (uint32_t)LV_PART_ANY | (uint32_t)LV_STATE_ANY);
 }
 
 /**
